+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "Project Dashboard" extension will be documented in this file. It follows the [Keep a Changelog](http://keepachangelog.com/) recommendations.
 
+## [2.0.1] 2020-04-122
+
+### Fixed
+    -   Context Menu is now correctly placed when the Dashboard is scrolled.
+
 ## [2.0.0] 2020-04-11
 
 ### Changed